+++ components/sftp/hooks/useSftpKeyboardShortcuts.ts
@@ -10,10 +10,10 @@ import type { MutableRefObject } from "react";
 import { KeyBinding, matchesKeyBinding } from "../../../domain/models";
 import { getParentPath, joinPath, resolveSftpWindowsPathOptions } from "../../../application/state/sftp/utils";
 import { netcattyBridge } from "../../../infrastructure/services/netcattyBridge";
-import { sftpClipboardStore, SftpClipboardFile } from "./useSftpClipboard";
-import { sftpFocusStore } from "./useSftpFocusedPane";
-import { sftpDialogActionStore } from "./useSftpDialogAction";
-import { sftpTreeSelectionStore } from "./useSftpTreeSelectionStore";
+import { sftpClipboardStore, SftpClipboardFile } from "../../../application/state/sftp/sftpClipboardStore";
+import { sftpFocusStore } from "../../../application/state/sftp/sftpFocusStore";
+import { sftpDialogActionStore } from "../../../application/state/sftp/sftpDialogActionStore";
+import { sftpTreeSelectionStore } from "../../../application/state/sftp/sftpTreeSelectionStore";
 import { sftpListOrderStore } from "./useSftpListOrderStore";
 import { sftpPaneViewModeStore } from "../../../application/state/sftp/sftpPaneViewModeStore";
 import { keepOnlyPaneSelections } from "./selectionScope";
