+++ components/terminal/useServerStats.test.ts
@@ -3,7 +3,7 @@ import { readFileSync } from "node:fs";
 import test from "node:test";
 
 test("server stats retry can restart polling after give-up", () => {
-  const source = readFileSync(new URL("./hooks/useServerStats.ts", import.meta.url), "utf8");
+  const source = readFileSync(new URL("../../application/state/useServerStats.ts", import.meta.url), "utf8");
   const reconcileStart = source.indexOf("function reconcileSharedServerStatsSession");
   const shouldRestartCheck = source.indexOf("const shouldRestartPolling", reconcileStart);
   const givenUpCheck = source.indexOf("if (session.givenUp) return;", shouldRestartCheck);
@@ -17,7 +17,7 @@ test("server stats retry can restart polling after give-up", () => {
 });
 
 test("server stats stale requests do not block a fresh visible fetch", () => {
-  const source = readFileSync(new URL("./hooks/useServerStats.ts", import.meta.url), "utf8");
+  const source = readFileSync(new URL("../../application/state/useServerStats.ts", import.meta.url), "utf8");
 
   assert.match(source, /inflightGeneration: number \| null/);
   assert.match(source, /session\.inflight && session\.inflightGeneration === session\.fetchGeneration/);
@@ -25,7 +25,7 @@ test("server stats stale requests do not block a fresh visible fetch", () => {
 });
 
 test("server stats polling does not stop when its terminal is in the background", () => {
-  const source = readFileSync(new URL("./hooks/useServerStats.ts", import.meta.url), "utf8");
+  const source = readFileSync(new URL("../../application/state/useServerStats.ts", import.meta.url), "utf8");
 
   assert.doesNotMatch(source, /isVisible/);
   assert.doesNotMatch(source, /getVisibleServerStatsClients/);
